Add simple test problems
On test_read_envvar_from_context_without_set_environment the developer doesn't set their envvar on test, then this test is dependent of local configuration files of user or default configuration (on settings.py), this is an wrong test.

On test_read_envvar_from_context_setting_correct_environment the developer set their envvar on test using settings fixture (from pytest-django), then this test is not dependent of local configuration files of user or default configuration (on settings.py), this is an correct test.